Progress on latest spec update
