+++ lang/en/lang.php
@@ -171,7 +171,7 @@
     'property_comment_position' => 'Comment Position',
     'property_comment_position_above' => 'Above',
     'property_comment_position_below' => 'Below',
-    'property_hint_path' => 'Hint partial path',
+    'property_hint_path' => 'Partial Path',
     'property_hint_path_description' => 'Path to a partial file that contains the hint text. Use the $ symbol to refer the plugins root directory, for example: $/acme/blog/partials/_hint.php',
     'property_hint_path_required' => 'Please enter the hint partial path',
     'property_partial_path' => 'Partial path',
@@ -313,16 +313,19 @@
     'property_switch_label_off_description' => 'Set a custom label for an "OFF" switch state',
     'control_group_standard' => 'Standard',
     'control_group_widgets' => 'Widgets',
+    'control_group_ui' => 'UI',
     'click_to_add_control' => 'Add control',
     'loading' => 'Loading...',
     'control_text' => 'Text',
     'control_text_description' => 'Single line text box',
+    'control_email' => 'Email',
+    'control_email_description' => 'Single line text box that takes email addresses only',
     'control_password' => 'Password',
     'control_password_description' => 'Single line password text field',
     'control_checkbox' => 'Checkbox',
     'control_checkbox_description' => 'Single checkbox',
     'control_switch' => 'Switch',
-    'control_switch_description' => 'Single switchbox, an alternative for checkbox',
+    'control_switch_description' => 'Single light switch input, an alternative for checkbox',
     'control_textarea' => 'Text Area',
     'control_textarea_description' => 'Multiline text box with controllable height',
     'control_dropdown' => 'Dropdown',
@@ -340,6 +343,8 @@
     'control_partial_description' => 'Outputs a partial contents',
     'control_section' => 'Section',
     'control_section_description' => 'Displays a form section with heading and subheading',
+    'control_ruler' => 'Horizontal Rule',
+    'control_ruler_description' => 'Displays a a horizontal rule to break up the contents',
     'control_radio' => 'Radio List',
     'control_radio_description' => 'A list of radio options, where only one item can be selected at a time',
     'control_radio_option_1' => 'Option 1',
