+++ widgets/defaultcontroldesigntimeprovider/partials/_control-email.php
@@ -1,0 +1,3 @@
+<div class="builder-blueprint-control-text">
+    <i class="icon-envelope"></i><?= e(trans('rainlab.builder::lang.form.control_email')) ?>
+</div>
