+++ widgets/defaultcontroldesigntimeprovider/partials/_control-ruler.php
@@ -1,0 +1,1 @@
+<div class="builder-blueprint-control-section pt-4"></div>
